sw-elad-s1

+++ devices/sw-elad-s1/elad-worker.cpp
@@ -2,25 +2,22 @@
 /*
  *    Copyright (C) 2014
  *    Jan van Katwijk (J.vanKatwijk@gmail.com)
- *    Lazy Chair Programming
+ *    Lazy Chair Computing
  *
- *    This file is part of the SDR-J.
- *    Many of the ideas as implemented in SDR-J are derived from
- *    other work, made available through the GNU general Public License. 
- *    All copyrights of the original authors are recognized.
+ *    This file is part of the spectrumViewer
  *
- *    SDR-J is free software; you can redistribute it and/or modify
+ *    spectrumViewer is free software; you can redistribute it and/or modify
  *    it under the terms of the GNU General Public License as published by
  *    the Free Software Foundation; either version 2 of the License, or
  *    (at your option) any later version.
  *
- *    SDR-J is distributed in the hope that it will be useful,
+ *    spectrumViewer is distributed in the hope that it will be useful,
  *    but WITHOUT ANY WARRANTY; without even the implied warranty of
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  *    GNU General Public License for more details.
  *
  *    You should have received a copy of the GNU General Public License
- *    along with SDR-J; if not, write to the Free Software
+ *    along with spectrumViewer; if not, write to the Free Software
  *    Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
@@ -112,15 +109,15 @@ int	rc;
 	fprintf (stderr, "eladWorker now stopped\n");
 }
 
-void	eladWorker::setVFOFrequency	(int32_t f) {
+void	eladWorker::setVFOFrequency	(uint64_t f) {
 	if (!runnable)
 	   return;
 //	fprintf (stderr, "Setting freq to %d\n", f);
-	lastFrequency	= f;
+	lastFrequency	= (long int)f;
 	functions	-> SetHWLO (functions -> getHandle (), &lastFrequency);
 }
 
-int32_t	eladWorker::getVFOFrequency	(void) {
-	return lastFrequency;
+uint64_t	eladWorker::getVFOFrequency	(void) {
+	return  (uint64_t)lastFrequency;
 }
